pyproject.toml
[project]
name = "spa-app"
version = "0.1.0"
description = "Dzen test"
authors = [
    {name = "AndriyIshchenko",email = "iscenkoa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
